moved all contents of git_utils.py to utils/git.py
